follow up bc9190ca3c5746728c15d7c1331d7cb4b3a25763: fix fallout from automerge
