+++ compiler-cli/integration-tests/semantic/main_method_redef_param.mj.stderr
@@ -1,5 +1,5 @@
-error: redefinition of local var 'args'
+error: redefinition of parameter 'args'
    | 
  3 |         int args = 42;
-   |             ^^^^
+   |         ^^^^^^^^^^^^^^
