+++ compiler-cli/integration-tests/semantic/this_method_invoc_inside_main.mj.stderr
@@ -1,7 +1,7 @@
 error: non-static method 'foo' cannot be referenced from a static context
    | 
  5 |         foo();
-   |         ^^^
+   |         ^^^^^
 
 error: non-static variable 'this' cannot be referenced from a static context
    |
